Switch to use com.zhokhov package. Need for Maven Central.
