+++ README.md
@@ -41,7 +41,7 @@ Add folowing to your `pom.xml`:
 
 ```xml
 <dependency>
-  <groupId>com.graphql-java</groupId>
+  <groupId>com.zhokhov.graphql</groupId>
   <artifactId>graphql-datetime-spring-boot-starter</artifactId>
   <version>2017-07-15T18-08-56</version>
 </dependency>
@@ -52,7 +52,7 @@ Add folowing to your `pom.xml`:
 Add folowing to your `build.gradle`:
 
 ```groovy
-compile 'com.graphql-java:graphql-datetime-spring-boot-starter:2017-07-15T18-08-56'
+compile 'com.zhokhov.graphql:graphql-datetime-spring-boot-starter:2017-07-15T18-08-56'
 ```
 
 ### Scalars
