+++ graphql-datetime-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-graphql.datetime.boot.GraphQLDateTimeAutoConfiguration
+com.zhokhov.graphql.datetime.boot.GraphQLDateTimeAutoConfiguration
