--- graphql-datetime-autoconfigure/src/test/java/graphql/datetime/boot/test/resolvers/Query.java
@@ -1,31 +1,0 @@
-/*
- *  Copyright 2017 Alexey Zhokhov
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package graphql.datetime.boot.test.resolvers;
-
-import com.coxautodev.graphql.tools.GraphQLRootResolver;
-import org.springframework.stereotype.Component;
-
-/**
- * @author <a href='mailto:alexey@zhokhov.com'>Alexey Zhokhov</a>
- */
-@Component
-public class Query implements GraphQLRootResolver {
-
-    public ResponseType echo() {
-        return new ResponseType();
-    }
-
-}
