+++ graphql-datetime-sample-app/src/main/java/com/zhokhov/graphql/datetime/sample/GraphqlDatetimeSampleAppApplication.java
@@ -1,0 +1,31 @@
+/*
+ *  Copyright 2017 Alexey Zhokhov
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package com.zhokhov.graphql.datetime.sample;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author <a href='mailto:alexey@zhokhov.com'>Alexey Zhokhov</a>
+ */
+@SpringBootApplication
+public class GraphqlDatetimeSampleAppApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GraphqlDatetimeSampleAppApplication.class, args);
+    }
+
+}
