--- graphql-datetime-sample-app/src/test/java/graphql/datetime/sample/GraphqlDatetimeSampleAppApplicationTests.java
@@ -1,34 +1,0 @@
-/*
- *  Copyright 2017 Alexey Zhokhov
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package graphql.datetime.sample;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-/**
- * @author <a href='mailto:alexey@zhokhov.com'>Alexey Zhokhov</a>
- */
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class GraphqlDatetimeSampleAppApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
