+++ graphql-java-datetime/src/main/java/com/zhokhov/graphql/datetime/DateTimeHelper.java
@@ -1,0 +1,119 @@
+/*
+ *  Copyright 2017 Alexey Zhokhov
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package com.zhokhov.graphql.datetime;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Date;
+import java.util.Objects;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.concurrent.TimeUnit;
+
+/**
+ * @author <a href='mailto:alexey@zhokhov.com'>Alexey Zhokhov</a>
+ */
+public class DateTimeHelper {
+
+    public static final CopyOnWriteArrayList<DateTimeFormatter> DATE_FORMATTERS = new CopyOnWriteArrayList<>();
+
+    static {
+        DATE_FORMATTERS.add(DateTimeFormatter.ISO_INSTANT.withZone(ZoneOffset.UTC));
+        DATE_FORMATTERS.add(DateTimeFormatter.ISO_LOCAL_DATE_TIME.withZone(ZoneOffset.UTC));
+        DATE_FORMATTERS.add(DateTimeFormatter.ISO_LOCAL_DATE.withZone(ZoneOffset.UTC));
+    }
+
+    // ISO_8601
+    public static String toISOString(LocalDateTime dateTime) {
+        Objects.requireNonNull(dateTime, "dateTime");
+
+        return DateTimeFormatter.ISO_INSTANT.format(ZonedDateTime.of(dateTime, ZoneOffset.UTC));
+    }
+
+    public static String toISOString(LocalDate date) {
+        Objects.requireNonNull(date, "date");
+
+        return DateTimeFormatter.ISO_LOCAL_DATE.format(date);
+    }
+
+    public static String toISOString(LocalTime time) {
+        Objects.requireNonNull(time, "time");
+
+        return DateTimeFormatter.ISO_LOCAL_TIME.format(time);
+    }
+
+    public static String toISOString(Date date) {
+        Objects.requireNonNull(date, "date");
+
+        return toISOString(toLocalDateTime(date));
+    }
+
+    public static LocalDateTime toLocalDateTime(Date date) {
+        Objects.requireNonNull(date, "date");
+
+        return date.toInstant().atZone(ZoneOffset.UTC).toLocalDateTime();
+    }
+
+    public static Date toDate(LocalDate date) {
+        Objects.requireNonNull(date, "date");
+
+        return toDate(date.atStartOfDay());
+    }
+
+    public static Date toDate(LocalDateTime dateTime) {
+        Objects.requireNonNull(dateTime, "dateTime");
+
+        return Date.from(dateTime.atZone(ZoneOffset.UTC).toInstant());
+    }
+
+    public static LocalDateTime parseDate(String date) {
+        Objects.requireNonNull(date, "date");
+
+        for (DateTimeFormatter formatter : DATE_FORMATTERS) {
+            try {
+                // equals ISO_LOCAL_DATE
+                if (formatter.equals(DATE_FORMATTERS.get(2))) {
+                    LocalDate localDate = LocalDate.parse(date, formatter);
+
+                    return localDate.atStartOfDay();
+                } else {
+                    return LocalDateTime.parse(date, formatter);
+                }
+            } catch (java.time.format.DateTimeParseException ignored) {
+            }
+        }
+
+        return null;
+    }
+
+    public static Date createDate(int year, int month, int day) {
+        return createDate(year, month, day, 0, 0, 0, 0);
+    }
+
+    public static Date createDate(int year, int month, int day, int hours, int min, int sec) {
+        return createDate(year, month, day, hours, min, sec, 0);
+    }
+
+    public static Date createDate(int year, int month, int day, int hours, int min, int sec, int millis) {
+        long nanos = TimeUnit.MILLISECONDS.toNanos(millis);
+        LocalDateTime localDateTime = LocalDateTime.of(year, month, day, hours, min, sec, (int) nanos);
+        return DateTimeHelper.toDate(localDateTime);
+    }
+
+}
